[tt-crank] feat(torch): init compile backend
Basic compile backend infrastructure is in place. The goal is to have
the simplest starting point possible - changes to follow will increase
supported ops count, add proper cpu fallback mechanism, etc.

The eager kernels are refactored so that we extract building of the TTIR
into a separate function - so that the same logic can be used by the
compile backend.

The compile backend gets the functionalized FX graph from the AOT
autograd - `aot_module_simplified`. Then, our FX interpreter constructs
the TTIR by calling IR builder functions - op-by-op.

After we have the complete IR, compiler pipeline is called and we get
the `CompiledProgram`. On forward, we will simply execute the flatbuffer
with the provided inputs.

There are some complications with `dtype` and scalars:
- model outputs could be expected as `int64`/`float64` - but our runtime
  always narrows tensors to 32-bit. To provide outputs as user would
  expect them we set torch metadata to be 64-bit `dtype`, and we will
  convert it accordingly when user calls `.to(cpu)`. This is the same
  hack that eager uses. To get the actual model outputs `dtype`s we use
  the `fx_node.meta["val"].dtype` from the FX graph
- scalar inputs are by default `int64`/`float64` - we convert them to
  tensors of type extracted from the FX graph:
  `fx_node.meta["val"].dtype`.

+++ tt-crank/src/torch/ops/linear.cpp
@@ -6,9 +6,9 @@
 #include <ttmlir/Dialect/TTIR/IR/TTIROps.h>
 
 #include "torch/backend.hpp"
+#include "torch/ops/builders.hpp"
 #include "torch/tensor.hpp"
 #include "torch/ttir_module_builder.hpp"
-#include "utils.hpp"
 
 namespace tt::kurbla::torch_backend {
 
@@ -53,10 +53,10 @@ at::Tensor tt_addmm(const at::Tensor &self, const at::Tensor &mat1, const at::Te
     } else {
         result = mb.create<mlir::tt::ttir::MatmulOp>(result_type, a, b, false, false).getResult();
         if (alpha.toDouble() != 1.0) {
-            result = scale_tensor(mb, result, alpha);
+            result = scale_tensor(mb, result, alpha.toDouble());
         }
         if (beta.toDouble() != 0.0) {
-            mlir::Value scaled_bias = beta.toDouble() == 1.0 ? bias : scale_tensor(mb, bias, beta);
+            mlir::Value scaled_bias = beta.toDouble() == 1.0 ? bias : scale_tensor(mb, bias, beta.toDouble());
             result = mb.create<mlir::tt::ttir::AddOp>(result_type, result, scaled_bias).getResult();
         }
     }
